+++ apps/sendmail/admin.py
@@ -2,16 +2,42 @@
 
 from django.contrib import admin
 
-from .models import Mail
+from .models import Mail, MailRecipient
 
 
 class MailAdmin(admin.ModelAdmin):
 
-    list_display = ['subject', 'sent_time', 'recipients_total', 'successful_mails', 'failed_mails', 'done_sending']
+    list_display = ['uid', 'subject', 'created_time', 'is_sending_complete']
+    ordering = ['-created_time']
+
+    # Prevent creation
+    def has_add_permission(self, request, obj=None):
+        return False
+
+    # Prevent deletion
+    def has_delete_permission(self, request, obj=None):
+        return False
+
+    # Prevent changes
+    def save_model(self, request, obj, form, change):
+        pass
+
+    # Prevent M2M changes
+    def save_related(self, request, form, formsets, change):
+        pass
+
+
+class MailRecipientAdmin(admin.ModelAdmin):
+
+    list_display = ['mail', 'user', 'sent_time']
     ordering = ['-sent_time']
 
     # Prevent creation
     def has_add_permission(self, request, obj=None):
+        return False
+
+    # Prevent deletion
+    def has_delete_permission(self, request, obj=None):
         return False
 
     # Prevent changes
@@ -24,3 +50,4 @@ class MailAdmin(admin.ModelAdmin):
 
 
 admin.site.register(Mail, MailAdmin)
+admin.site.register(MailRecipient, MailRecipientAdmin)