+++ apps/userprofile/forms.py
@@ -12,13 +12,14 @@ from apps.userprofile.models import UserProfile
 class UserProfileForm(forms.ModelForm):
     class Meta:
         model = UserProfile
-        fields = ('nick', 'date_of_birth', 'address', 'zip_code', 'phone')
+        fields = ('nick', 'date_of_birth', 'address', 'zip_code', 'phone', 'marketing_optin')
         widgets = {
             'nick': forms.TextInput(attrs={'class': 'form-control', 'placeholder': _(u'Nickname'), 'type': 'text'}),
             'date_of_birth': forms.TextInput(attrs={'class': 'form-control', 'type': 'date'}),
             'address': forms.TextInput(attrs={'class': 'form-control', 'placeholder': _(u'Street address'), 'type': 'text'}),
             'zip_code': forms.TextInput(attrs={'class': 'form-control', 'placeholder': _(u'Postal code'), 'type': 'number'}),
-            'phone': forms.TextInput(attrs={'class': 'form-control', 'placeholder': _(u'Phone number'), 'type': 'tel'})}
+            'phone': forms.TextInput(attrs={'class': 'form-control', 'placeholder': _(u'Phone number'), 'type': 'tel'}),
+            'marketing_optin': forms.CheckboxInput()}
 
     def clean(self):
         cleaned_data = super(UserProfileForm, self).clean()