+++ CHANGELOG.md
@@ -9,7 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added marketing email opt-in for users.
+
 ### Changed
+
+- Reworked custom email sending.
 
 ### Deprecated